refactor(llm/AnalyticsProvider): unused provider
